requirements.txt
pandas==2.2.2
Pillow==10.3.0
websocket_client==1.8.0
